Become asynchronous - use Motor and AsyncIO
Motor is the asynchronous MongoDB driver (PyMongo being the synchronous
equivalent).
Motor has support for AsyncIO.
https://motor.readthedocs.io/en/stable/index.html

A new async MongoCollection has been created and used to finalize the
/gateways-endpoint concerning POST requests.
This has led to minor updates of pydantic models changes.

The server also includes the optimade package middleware and exceptions,
as well as the exposing the current routers (gateways and structures).