Depend on pydantic and fastapi through optimade
Since we will be using core parts of the `optimade` package, we wish to
not clash with its version dependencies on pydantic and fastapi.
So to ensure the novel pip resolver is happy and we have a stable
package, we will install and update pydantic and fastapi according to
the requirements of the `optimade` package.