First working GET /gateways/{id}/structures
Very course first implementation of the GET
/gateways/{gateway_id}/structures endpoint, including a test.

In order to start the external DB requests simultaneously,
concurrent.futures.ThreadPoolExecutor() is used according to the example
https://docs.python.org/3/library/concurrent.futures.html#threadpoolexecutor-example

The main issue now is how to handle pagination (gateway versus
databases). For now, the implementation simply passes through all
parameters unchanged to the databases and ensure the "next" link will
get the correct next resources from the databases, but at the cost of
the original `page_limit` not being respected in the number of results
that are shown on a single page by the gateway.
I.e., if `page_limit=5` for a two database gateway (with databases of
>5 entries), the resulting standard GET request will return 10 entries,
5 from each of the databases.