Remove collection caching
An issue might occur, where the cached/stored data_returned or
data_available value is invalid, if, e.g., multiple queries are
performed simultaneously.

This removes all forms of caching, which was only relevant for the
data_returned and data_available properties. Hence they have also been
removed as properties.

The count() method is used instead, now with support for the
EntryListingQueryParams/SingleEntryQueryParams parameters, as well as
passing a dictionary of keyword arguments of valid Mongo counting
parameters.