Implement querying as a background concurrent task
A `perform_query` function is implemented that uses
`asyncio.as_completed()` to retrieve a list of `asyncio.Task`s that each
perform database GET requests to a gateway's databases.

The QueryResource model has been updated to reflect the information
needed and the AsyncMongoCollection.create_one has gotten a `set_id`
functionality in order to make sure a `QueryResource` gets an "id" field
in the MongoDB.

Instead of adding `perform_query` as a background task, it can be set up
to use an `asyncio.Queue` or similar, but for now this will work.
To get more information about the background task system built into
FastAPI see here:
https://fastapi.tiangolo.com/tutorial/background-tasks/