Use singledb gateway for CI - update id filtering
Due to the gateways changing each entry's `id` value, the filtering has
to be updated for each database.
This implements a regex to find the `id` values (both id <operator>
value and value <operator> id).
The values are then updated specifically for each database-specific
query only, and kept as is in the
QueryResource.attributes.query_parameters.filter value.