Update queries API - add prepare.py
Set up prepare.py and perform.py.
Move most of the preparations needed to start the query tasks in
`perform_query` to separate functions in prepare.py.
Collect the needed preparation functions in a single `prepare_query`
function that can be used in `perform_query` to retrieve all necessary
variables.