Test and update creating queries
Make standard status code response 202 Accepted for POST requests.

Implement general create resource model to include the possibility of
setting the `id`.

Only have a single "finished" query state.
There is no need to have the split of "errored" and "successful".
If these were there it should also not be related to the response, but
rather whether the query for some reason (from the side of the gateway
service) failed/errored. This is not something to be implemented at this
point, as a special error handler needs to be implemented for this to
update the database entry.

Use static local data for tests.
Utilize the `add_callback` feature of pytest_httpx.
If a database is now ID'd as `sleep-X`, where X is an integer, the
`mock_responses` fixture will now run a function that sleeps for X
seconds before returning the standard response for `optimade-sample`.

Use flaky for query_results test as it fails 1/3 of the time.