Use OPTIMADE Python tools test server for CI
The OPTIMADE Python tools docker image is run for the docker CI job and
added as a database to a new gateway.

The tests are updated to dynamically test according to the loaded test
data and not hard coded.

Connect OPT server to gateway network and use the docker-local port
(5000).