Don't use BackgroundTask - use ThreadPoolExecutor
To ensure a responsive gateway API - as well as easier/more intuitive
tests, BackgroundTask is not used - it blocks the OS thread, where the
asyncio event loop is running.
In order to avoid this, a ThreadPoolExecutor (from concurrent.futures)
is used instead and the calls to `perform_query()` is done via
`asyncio.create_task()`.

With this, the slow query responses can be mocked. To this end the
actual sleep is done in a ThreadPoolExecutor as well to not block the
event loop used when running tests - or more specifically, to not block
the thread in which the event loop is running.

Document ThreadPoolExecutor design decision in docs/design.md.

Include several other minor fixes, the most major being changing the
return status code for POST /gateways from 202 to 200.