Test special gateway IDs - new Exception handler
Gateway IDs cannot be URL paths, i.e., they cannot contain forward
slashes (`/`).
This is ensured by defining a regex for `id` in `GatewayResource` and
`GatewayCreate`.

Furthermore, the resulting `RequestValidationError` raised if this is
wrong is captured by a new exception handler that's used instead of the
one from `optimade`.
This will make the HTTP status code 400 if it's a POST method, since it
will be the posted body that's wrong - i.e. a client error.
Furthermore, it will imitate the `ValidationError` exception handler
from `optimade` and run through all `errors` (attribute).