Add documentation through MkDocs (+ CD)
A new list of dependencies for a `docs` extra has been added to add
MkDocs support.
The MkDocs implementation is based on the same one for OPTIMADE Python
tools.

An invoke task is added to create the API reference documentation as
well as write the landing page (`index.md`) from `README.md`. The main
difference is some relative link target alignment.

An empty `CHANGELOG.md` is added as a placeholder for a future automated
changelog generation.

Add CI job to test documentation build.

Add GH Actions CD for documentation:
The CD workflow will first update the CHANGLOG.md file, then update the
documentation, commit and push the changes to `main` on GitHub.
Then the release description/body will be updated with the
release-specific changelog.
Finally, another job will deploy the documentation to the `gh-pages`
branch.

There is a commented out job for publishing to PyPI, which is ready to
be implemented once this is relevant.

Dependabot is updated with relevant labels.

The files in the `.github` folder is reordered to be more conceptually
clean.