New query endpoint response model
Instead of returning a `GatewayQueryResponse` for the single
`QueryResource` endpoints as well as the `GET /search` endpoint, a
`QueriesResponseSingle` is now returned.
This is done to ensure all necessary information is communicated along
with the actual data response, like the used query parameters and such.

The example/test gateways have been updated to use "correct" database
`id`s, where the provider is prepended, separating the provider `id` and
the database `id` with a slash.