Adapt to updated optimade
Remove root validator from `Response` for `GatewayQueryResponse`.
The `GatewayQueryResponse` is special, since it can allow `errors` to be
present together with `data`.
And `meta` is guaranteed to always be present, so there is no need for
the validator.

Also, update the test data to respect the new validators for
`chemical_formula_anonymous`.

Update optimade and pre-commit dependencies.