Satisfy pylint + collection_factory
In order to avoid cyclic imports and clean up the code, an
entry-endpoint resource collection factory has been implemented.
It updates a global dictionary, where the key is the MongoDB collection
name configured and retrieved from `CONFIG`. The value is then an
asynchronous MongoDB collection similar to the `MongoCollection` class
in OPTIMADE Python tools.