Use `bandit`, `pylint`, `safety`, `mypy`, and `pyproject.toml` - collection factory (#119)
Add/remove pre-commit hooks:
Remove: `flake8`
Add: `bandit`, `pylint`, `mypy`, `pytest-plugins` (local)
Update `black` hook to the latest version.

Run and satisfy the new tools.

Use `typing.TYPE_CHECKING` throughout.
Since FastAPI depends on pydantic and relies on the type definitions to
function, the routers cannot use `TYPE_CHECKING`. However, it is used
to the extent it can be done.

Create a factory for entry-endpoint resource collections.
In order to avoid cyclic imports and clean up the code, an
entry-endpoint resource collection factory has been implemented.
It updates a global dictionary, where the key is the MongoDB collection
name configured and retrieved from `CONFIG`. The value is then an
asynchronous MongoDB collection similar to the `MongoCollection` class
in OPTIMADE Python tools.
With this, the `*_COLLECTIONS` can be removed throughout as well.

Add new CI job for `pylint` and `safety`.
`pylint` relies on having access to imported packages.
It is therefore skipped in the `pre-commit` job and given its own.

Make sure to import types for building docs.
This is done by setting the environment variable `MKDOCS_BUILD`
when building the docs. It is discarded after the build.

Don't use doc-strings through from `optimade`:
There was an issue with this, where a local API reference was made,
which couldn't be resolved when building these docs.

Use `pyproject.toml` throughout (remove `setup.cfg`).

Add `invoke` task to update pytest config plugins.
The task updates the `required_plugins` config option for `pytest` with
the versions specified in `requirements_dev.txt`.
It is implemented as a `pre-commit` hook as well with the id
`pytest-plugins`.