Combine workflows
Since the documentation deployment workflow for new updates to the
`main` branch starts on the same condition as the keep up-to-date
workflow, these have been combined in a single workflow as two separate
jobs.

Updated the workflow names and filenames, updated the badge in the
README as well in this case.