Follow redirects & Use httpx.Response throughout
The new version 0.20.0 of httpx sets `follow_redirects=False` by
default, this is now set to `True` in the `client` fixture.

Use `httpx.Response` instead of `pytest_httpx.to_response`, since the
latter is deprecated with the new version of `pytest_httpx`, due to
using `httpx.Response` directly in that function instead of
`httpcore.Response`.